moved the example from site to storybook
