+++ packages/embla-carousel/stories/carousel.stories.tsx
@@ -10,6 +10,7 @@ import {
 import type { Meta, StoryFn } from "@storybook/react-vite";
 import "./carousel.stories.css";
 import {
+  Avatar,
   Display1,
   FlexLayout,
   H2,
@@ -151,6 +152,65 @@ export const MultiSlide: StoryFn<typeof Carousel> = (args) => {
               >
                 <Text>{slide.content}</Text>
               </CarouselCard>
+            );
+          })}
+        </CarouselSlides>
+      </StackLayout>
+    </Carousel>
+  );
+};
+
+const avatarSlides = [
+  { name: "Alex Brown", color: "accent" },
+  { name: "Priya Patel", color: "category-1" },
+  { name: "Chen Wei", color: "category-2" },
+  { name: "Maria Silva", color: "category-3" },
+  { name: "Sam Taylor", color: "category-4" },
+  { name: "Emma Wilson", color: "category-5" },
+  { name: "Liam O'Neill", color: "category-6" },
+  { name: "Fatima Khan", color: "category-7" },
+  { name: "Noah Cohen", color: "category-8" },
+  { name: "Ava Bennett", color: "category-9" },
+  { name: "Diego Morales", color: "category-10" },
+  { name: "Zara Ahmed", color: "category-11" },
+  { name: "Ethan Clark", color: "category-12" },
+  { name: "Sofia Bailey", color: "category-13" },
+  { name: "Oliver Hayes", color: "category-14" },
+] as const;
+
+export const CustomSlide: StoryFn<typeof Carousel> = (args) => {
+  const carouselId = useId();
+  return (
+    <Carousel
+      aria-labelledby={`${carouselId}-title`}
+      className="avatarCarousel"
+      emblaOptions={{ align: "start", slidesToScroll: "auto" }}
+      {...args}
+    >
+      <H2 id={`${carouselId}-title`} className="carouselHeading">
+        Carousel example with avatar as slides
+      </H2>
+      <StackLayout gap={1} direction="column-reverse">
+        <FlexLayout gap={1} wrap={true} align={"center"}>
+          <CarouselPreviousButton aria-label="Previous avatar group" />
+          <CarouselNextButton aria-label="Next avatar group" />
+          <CarouselProgressLabel />
+        </FlexLayout>
+        <CarouselSlides>
+          {avatarSlides.map((avatar, index) => {
+            const slideId = `${carouselId}-slide${index}`;
+            return (
+              <div
+                aria-label={avatar.name}
+                role="group"
+                aria-roledescription="slide"
+                className="avatarSlide"
+                key={slideId}
+              >
+                <span className="avatarSlideIndicator" aria-hidden>
+                  <Avatar name={avatar.name} color={avatar.color} size={2} />
+                </span>
+              </div>
             );
           })}
         </CarouselSlides>
