+++ site/docs/components/carousel/examples.mdx
@@ -82,9 +82,3 @@ The Embla API enables programmatic control, such as the visible slide.
 </Callout>
 
 <LivePreview componentName="carousel" exampleName="ControlledCarousel" />
-
-## Custom slide
-
-You can render fully custom slide content, such as `Avatar` components, by composing your own slide elements inside `CarouselSlides`.
-
-<LivePreview componentName="carousel" exampleName="CustomCarouselSlide" />
